Backend/requirements.txt
django
django-cors-headers
djangorestframework
gunicorn